refactor: move SubjectUtil to kyuubi-util; fix Javadoc and Kyuubi-managed comments
